LoveSimulation_plan/requirements.txt
python-pptx
markdown
requests
